refactor app shell. update primary color
