+++ packages/frontend/app/search/components/index.ts
@@ -1,5 +1,6 @@
-export { default as Filters } from "./filters/filters";
-export { default as Pagination } from "./pagination";
+export { default as Filters, type FiltersProps } from "./filters/filters";
+export { default as Pagination, type PaginationProps } from "./pagination";
 export { default as AppShell } from "./appShell/appShell";
+export { default as AppShellClient } from "./appShell/appShellClient";
 export { default as CheckboxGroup } from "./checkboxGroup/checkboxGroup";
 export { default as Multiselect } from "./multiselect/multiselect";
