+++ packages/frontend/app/search/page.tsx
@@ -1,8 +1,8 @@
 import { Text } from "@mantine/core";
-import { fetchRecipeByQuery } from "./actions";
+import { fetchRecipeByQuery, fetchFilters } from "./actions";
 import { Card } from "@/components";
 import { RecipeSearchParams } from "@/types";
-import { AppShell, Pagination, Filters } from "./components";
+import { AppShellClient } from "./components";
 
 interface SearchPageProps {
   searchParams: Promise<RecipeSearchParams>;
@@ -11,15 +11,16 @@ interface SearchPageProps {
 export default async function SearchPage({ searchParams }: SearchPageProps) {
   const params = await searchParams;
   const { recipes, total } = await fetchRecipeByQuery(params);
+  const filterProps = await fetchFilters();
 
   if (!recipes.length) {
     return <Text>No recipes found</Text>;
   }
   return (
-    <AppShell filters={<Filters />} pagination={<Pagination total={total} />}>
+    <AppShellClient paginationProps={{ total }} filterProps={filterProps}>
       {recipes.map((recipe) => (
         <Card key={recipe.id} recipe={recipe} />
       ))}
-    </AppShell>
+    </AppShellClient>
   );
 }
